fix: host-specific item can sometimes be populated twice to destination
When a non-host-specific item `A` exists along with an host-specific
item `A@@hostname` for current host machine and both are specified as
sources in the same group, `A` will be populated first to its
destination, then `A@@hostname` will overwrite the just-written
destination.  This did not affect the final result but this behaviour
was unexpected.

Signed-off-by: Gaoyang Zhang <gy@blurgy.xyz>